refactor: index properties + index check fields helpers
