+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/constants.ts
@@ -8,7 +8,7 @@
 import { EcsFlat } from '@elastic/ecs';
 import { EuiComboBoxOptionOption } from '@elastic/eui';
 
-import { EcsFieldMetadata, SortConfig } from './types';
+import { EcsFieldMetadata, PartitionedFieldMetadata, SortConfig } from './types';
 import * as i18n from './translations';
 
 export const EcsFlatTyped = EcsFlat as unknown as Record<string, EcsFieldMetadata>;
@@ -48,4 +48,12 @@ export const defaultSort: SortConfig = {
     direction: 'desc',
     field: 'docsCount',
   },
+};
+
+export const EMPTY_METADATA: PartitionedFieldMetadata = {
+  all: [],
+  ecsCompliant: [],
+  custom: [],
+  incompatible: [],
+  sameFamily: [],
 };
