+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/indices_details/pattern/index_check_flyout/index_properties/index_check_fields/constants.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const ALL_TAB_ID = 'allTab';
+export const ECS_COMPLIANT_TAB_ID = 'ecsCompliantTab';
+export const CUSTOM_TAB_ID = 'customTab';
+export const INCOMPATIBLE_TAB_ID = 'incompatibleTab';
+export const SAME_FAMILY_TAB_ID = 'sameFamilyTab';
