+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/stub/get_check_state/index.ts
@@ -7,10 +7,6 @@
 
 import { IndicesGetMappingIndexMappingRecord } from '@elastic/elasticsearch/lib/api/types';
 
-import {
-  getMappingsProperties,
-  getSortedPartitionedFieldMetadata,
-} from '../../data_quality_details/indices_details/pattern/index_check_flyout/index_properties/helpers';
 import { mockMappingsResponse } from '../../mock/mappings_response/mock_mappings_response';
 import { UseIndicesCheckCheckState } from '../../hooks/use_indices_check/types';
 import { getUnallowedValues } from '../../utils/fetch_unallowed_values';
@@ -18,6 +14,7 @@ import { getUnallowedValueRequestItems } from '../../utils/get_unallowed_value_r
 import { EcsFlatTyped } from '../../constants';
 import { mockUnallowedValuesResponse } from '../../mock/unallowed_values/mock_unallowed_values';
 import { UnallowedValueSearchResult } from '../../types';
+import { getMappingsProperties, getSortedPartitionedFieldMetadata } from '../../utils/metadata';
 
 export const getCheckState = (
   indexName: string,
